test: add interface definitions for ExerciseDetails and Workout in tests
